refactor code:moved leads api functions in lead-service.tsx
