+++ services/lead-services.ts
@@ -1,12 +1,32 @@
-import { ApiService } from '@/services/api-services';
-import moment from 'moment';
+import { API_METHODS, API_PATHS } from '@/utils/constants/common-constants';
+import { AxiosRequestConfig } from 'axios';
+import { HttpClient, Response } from './axios-base-query';
 
 export class LeadService {
+  client;
+
+  constructor(baseUrl?: string) {
+    this.client = new HttpClient(baseUrl);
+  }
+
+   public getExecutives = async (token: string): Promise<any> => {
+    const config: AxiosRequestConfig = {};
+
+    if (token) {
+      config.headers = { Authorization: `Bearer ${token}` };
+    }
+    const resp = await this.client.request({
+      url: API_PATHS.GetExecutives,
+      method: API_METHODS.GET,
+      ...config,
+    });
+
+    return resp;
+  };
 
   public getExecutivesData = async (setExecutivesOption: any, token: string) => {
       try {
-        const ApiServices = new ApiService();
-        const response = await ApiServices.getExecutives(token);
+        const response = await this.getExecutives(token);
         const executivesOption = this.createSelectData(response.data.Data.Data);
         setExecutivesOption(executivesOption);
       } catch (error) {
@@ -25,10 +45,26 @@ export class LeadService {
     return selectOptions;
   };
 
+  public getLeads = async (token: string): Promise<any> => {
+    const config: AxiosRequestConfig = {};
+
+    if (token) {
+      config.headers = { Authorization: `Bearer ${token}` };
+    }
+
+    const resp = await this.client.request({
+      url: API_PATHS.GetLeads,
+      method: API_METHODS.GET,
+      ...config,
+    });
+
+    return resp;
+  };
+
   public getLeadsData = async (setLeadsData:any, token: string) => {
     try {
-      const ApiServices = new ApiService();
-      const response = await ApiServices.getLeads(token);
+
+      const response = await this.getLeads(token);
       console.log(response);
       const data = response.data.Data.Data;
       setLeadsData(data);
